Develop

+++ app/Http/Controllers/Admin/BrandController.php
@@ -3,10 +3,12 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Helpers\CollectionHelper;
-use App\Helpers\StringHelper;
+use App\Helpers\ResponseHelper;
 use App\Http\Controllers\Controller;
+use App\Repositories\Shop\Brand\BrandCreateRequest;
 use App\Repositories\Shop\Brand\BrandRepositoryInterface;
-use Illuminate\Validation\Rule;
+use App\Repositories\Shop\Brand\BrandUpdateRequest;
+use Illuminate\Database\QueryException;
 
 class BrandController extends Controller
 {
@@ -29,44 +31,31 @@ class BrandController extends Controller
         return $this->brandRepository->find($slug);
     }
 
-    public function store()
+    public function store(BrandCreateRequest $request)
     {
-        $brand = $this->brandRepository->create(self::validateCreate());
-        return response()->json($brand, 201);
+        try {
+            $brand = $this->brandRepository->create($request->validated());
+            return response()->json($brand, 201);
+        } catch (QueryException $e) {
+            return ResponseHelper::generateValidateError('code', 'The code has already been taken.');
+        }
     }
 
-    public function update($slug)
+    public function update(BrandUpdateRequest $request, $slug)
     {
-        return $this->brandRepository->update($slug, self::validateUpdate($slug));
+        try {
+            return $this->brandRepository->update($slug, $request->validated());
+        } catch (QueryException $e) {
+            return ResponseHelper::generateValidateError('code', 'The code has already been taken.');
+        }
     }
 
     public function destroy($slug)
     {
-        return response()->json(['message' => 'Permission denied.'], 403);
+        if ($this->brandRepository->checkProducts($slug)) {
+            return response()->json(['message' => 'Cannot delete brand if there is a linked product.'], 403);
+        }
 
-        $this->brandRepository->delete($slug);
-        return response()->json(['message' => 'successfully deleted'], 200);
-    }
-
-    private static function validateCreate()
-    {
-        request()->merge(['slug' => StringHelper::generateUniqueSlug()]);
-
-        return request()->validate([
-            'name' => 'required|unique:brands',
-            'slug' => 'required|unique:brands',
-            'image_slug' => 'nullable|exists:App\Models\File,slug',
-        ]);
-    }
-
-    private static function validateUpdate($slug)
-    {
-        return request()->validate([
-            'name' => [
-                'required',
-                Rule::unique('brands')->ignore($slug, 'slug'),
-            ],
-            'image_slug' => 'nullable|exists:App\Models\File,slug',
-        ]);
+        return $this->brandRepository->delete($slug);
     }
 }
